partially split the app into diffrent files
